refactor: Continue moving files and integrating the two server implementations
