--- src/aux-auth/nodemon.json
@@ -1,3 +1,0 @@
-{
-    "ignore": ["aux-web/*"]
-}
