--- src/aux-auth/typings/global.d.ts
@@ -1,4 +1,0 @@
-declare module '*.txt' {
-    const value: string;
-    export default value;
-}
