+++ src/aux-server/aux-backend/server/directory/DirectoryClientSettings.ts
@@ -17,10 +17,10 @@ export interface DirectoryClientSettings {
     /**
      * The JWT that the client last got from the server.
      */
-    token: string;
+    token: string | null;
 
     /**
      * The key of the client.
      */
-    key: string;
+    key: string | null;
 }
