+++ src/aux-server/aux-backend/server/directory/DirectoryEntry.ts
