+++ src/aux-server/aux-backend/server/modules/SerialModule.ts
@@ -44,7 +44,7 @@ export class SerialModule implements AuxModule2 {
         sub.add(
             simulation.localEvents
                 .pipe(
-                    flatMap(async event => {
+                    flatMap(async (event) => {
                         if (event.type === 'serial_connect') {
                             await this._serialConnect(simulation, event);
                         }
@@ -88,9 +88,23 @@ export class SerialModule implements AuxModule2 {
     _serialConnect(simulation: Simulation, event: SerialConnectAction) {
         try {
             // Complete the bluetooth connection before opening it up
-            let jsond = '{"command":"connect","device":"' + event.device + '", "mac":"' + event.mac + '", "channel":"' + event.channel + '"}';
-            let ip = execSync('ip route show | awk \'/default/ {print $3}\'').toString().trim();
-            let curl_command = 'curl -H "Content-Type: application/json" -X POST -d \'' + jsond + '\' ' + ip + ':8090/post';
+            let jsond =
+                '{"command":"connect","device":"' +
+                event.device +
+                '", "mac":"' +
+                event.mac +
+                '", "channel":"' +
+                event.channel +
+                '"}';
+            let ip = execSync("ip route show | awk '/default/ {print $3}'")
+                .toString()
+                .trim();
+            let curl_command =
+                'curl -H "Content-Type: application/json" -X POST -d \'' +
+                jsond +
+                "' " +
+                ip +
+                ':8090/post';
 
             execSync(curl_command);
 
@@ -273,9 +287,17 @@ export class SerialModule implements AuxModule2 {
     _serialClose(simulation: Simulation, event: SerialCloseAction) {
         try {
             // Send a command to kill the rfcomm process
-            let jsond = '{"command":"disconnect","device":"' + event.device + '"}';
-            let ip = execSync('ip route show | awk \'/default/ {print $3}\'').toString().trim();
-            let curl_command = 'curl -H "Content-Type: application/json" -X POST -d \'' + jsond + '\' ' + ip + ':8090/post';
+            let jsond =
+                '{"command":"disconnect","device":"' + event.device + '"}';
+            let ip = execSync("ip route show | awk '/default/ {print $3}'")
+                .toString()
+                .trim();
+            let curl_command =
+                'curl -H "Content-Type: application/json" -X POST -d \'' +
+                jsond +
+                "' " +
+                ip +
+                ':8090/post';
 
             execSync(curl_command);
