+++ src/aux-server/aux-backend/server/modules/SetupChannelModule2.ts
