+++ src/aux-server/aux-backend/server/player.config.ts
