+++ src/aux-server/package.json
@@ -5,16 +5,18 @@
     "main": "server/dist/main.js",
     "private": true,
     "scripts": {
-        "start": "node server/dist/main.js",
+        "start": "node aux-backend/server/dist/main.js",
         "watch": "concurrently --names \"Server,Player\" \"npm:watch:server\" \"npm:watch:player\"",
-        "watch:secure": "concurrently --names \"Server,Player\" \"npm:watch:server\" \"npm:watch:player:secure\"",
+        "watch:secure": "concurrently --names \"Server,Player,Auth\" \"npm:watch:server\" \"npm:watch:player:secure\"",
         "watch:server": "concurrently --names \"Build,Inst\" \"node ./script/dev-server.js\" \"npm:watch:inst\"",
         "watch:player": "vite serve ./aux-web/aux-player",
+        "watch:auth": "vite serve ./aux-web/aux-auth",
         "watch:player:secure": "vite serve ./aux-web/aux-player --host --https",
-        "watch:inst": "wait-on server/dist/main.js && nodemon server/dist/main.js",
+        "watch:auth:secure": "vite serve ./aux-web/aux-auth --host --https",
+        "watch:inst": "wait-on aux-backend/server/dist/main.js && nodemon aux-backend/server/dist/main.js",
         "vite": "vite serve ./aux-web/aux-player",
         "build": "npm run build:clean && npm run build:server && npm run build:web",
-        "build:web": "npm run build:interpreter && node --max_old_space_size=6144 ../../node_modules/vite/bin/vite.js build ./aux-web/aux-player",
+        "build:web": "npm run build:interpreter && node --max_old_space_size=6144 ../../node_modules/vite/bin/vite.js build ./aux-web/aux-player && node --max_old_space_size=6144 ../../node_modules/vite/bin/vite.js build ./aux-web/aux-auth",
         "build:server": "node --max_old_space_size=6144 ./script/build-server.js",
         "build:interpreter": "node --max_old_space_size=6144 ./script/build-interpreter.js",
         "build:docs": "echo \"Nothing to build...\"",
@@ -44,18 +46,20 @@
     "dependencies": {
         "@casual-simulation/aux-common": "^3.1.32",
         "@casual-simulation/aux-records": "^3.1.32",
+        "@casual-simulation/aux-records-aws": "^3.1.32",
         "@casual-simulation/aux-vm": "^3.1.32",
         "@casual-simulation/aux-vm-browser": "^3.1.32",
         "@casual-simulation/aux-vm-client": "^3.1.32",
         "@casual-simulation/aux-vm-deno": "^3.1.32",
         "@casual-simulation/aux-vm-node": "^3.1.32",
-        "@casual-simulation/casual-apiary-redis": "^3.1.32",
         "@casual-simulation/causal-tree-client-websocket": "^3.1.29",
         "@casual-simulation/causal-tree-server": "^3.1.29",
         "@casual-simulation/causal-tree-server-websocket": "^3.1.29",
         "@casual-simulation/causal-tree-store-cassandradb": "^3.1.29",
         "@casual-simulation/causal-tree-store-mongodb": "^3.1.29",
         "@casual-simulation/causal-trees": "^3.1.29",
+        "@casual-simulation/casual-apiary": "^3.1.32",
+        "@casual-simulation/casual-apiary-redis": "^3.1.32",
         "@casual-simulation/crypto": "^3.1.23",
         "@casual-simulation/crypto-browser": "^3.1.23",
         "@casual-simulation/crypto-node": "^3.1.23",
@@ -68,6 +72,7 @@
         "@casual-simulation/tunnel": "^3.1.28",
         "@casual-simulation/vue-shortkey": "^3.1.20",
         "@casual-simulation/websocket": "^3.1.11",
+        "@casual-simulation/rate-limit-redis": "^3.1.29",
         "@chenfengyuan/vue-qrcode": "^1.0.0",
         "@ericblade/quagga2": "1.4.2",
         "@hapi/joi": "15.1.1",
