--- src/aux-server/server/index.ts
@@ -1,57 +1,0 @@
-/**
- * MIT License
- *
- * Copyright (c) 2019 Casual Simulation, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- * @license MIT
- */
-
-import process from 'process';
-import { Server } from './server';
-
-import prodConfig from './config.prod';
-import devConfig from './config.dev';
-import { Config } from './config';
-
-const env = process.env.NODE_ENV;
-let config: Config;
-if (env === 'production') {
-    config = prodConfig;
-} else {
-    config = devConfig;
-}
-
-const server = new Server(config);
-
-async function init() {
-    await configure();
-    server.start();
-}
-
-init();
-
-async function configure() {
-    try {
-        await server.configure();
-    } catch (ex) {
-        console.error(ex);
-        process.exit(1);
-    }
-}
