--- src/aux-server/server/lb-services-config.js
@@ -1,4 +1,0 @@
-//client config - test
-const lbServicesConfig = {
-    url: 'https://filesimulator.com',
-};
